move sessioncache from internal
(cherry picked from commit 7051a13dbd92190258c5add5a1784ab0835f5163)
