+++ sessioncache/store_singlekritem.go
