+++ sessioncache/store_singlekritem_test.go
