+++ sessioncache/store_tester.go
